Resolved navigation issues
