+++ Escola em Casa/Controllers/AboutViewController.swift
@@ -6,7 +6,5 @@ class ThirdViewController: UIViewController {
         super.viewDidLoad()
         // Do any additional setup after loading the view.
     }
-
-
 }
